BACK Team details (re-added release player, fixed contract usage)

+++ util/dateFormatter.ts
@@ -7,7 +7,7 @@ const calculateAge = (dateOfBirth: Date): number => {
 
 const formatDate = (date: Date): string => {
     const year = new Intl.DateTimeFormat('en', { year: 'numeric' }).format(date);
-    const month = new Intl.DateTimeFormat('en', { month: 'short' }).format(date);
+    const month = new Intl.DateTimeFormat('en', { month: 'numeric' }).format(date);
     const day = new Intl.DateTimeFormat('en', { day: '2-digit' }).format(date);
 
     return `${day}/${month}/${year}`;
